backend/requirement.txt
flask
flask-cors
pymongo
bcrypt
PyJWT